refactor(layout): extract StudyLayout side effects into 4 dedicated hooks
StudyLayout is the central layout for the entire participant flow. It
hosts ten cooperating concerns inline (progress reporting, draft
auto-save with debounce + beforeunload + unmount flush, URL ?lang=
override, i18n / <html lang> sync, browser title, welcome-back toast,
click-outside menu management, stale session reset, scroll-to-top,
loader hydration). Cyclomatic complexity 74 on the function shell made
every change risky.

This refactor pulls the four most logic-dense concerns into named hooks
under frontend/src/hooks/study/. Each hook owns its effect, its state,
and its teardown, and is unit-testable independently of the layout's
JSX.

- useProgressReporter(slug)
  Subscribes to the session store and reports each step advance to
  /api/study/{slug}/progress. Skipped in pilot mode and without a
  session token. Failures silent — participant experience unaffected.
- useDraftAutoSave(slug)
  Debounced (5s) auto-save of presort/rough/qsort/postsort to
  /api/study/{slug}/save-draft. Flushes via keepalive fetch on
  beforeunload and via customInstance on unmount. Returns
  draftSaveStatus ('idle' | 'saving' | 'saved' | 'error') for the
  layout chrome.
- useStudyLocaleSync({config, sessionLanguage, hasConsented, ...})
  Coordinates four cooperating effects: URL ?lang= override (only when
  in available_languages), i18n + <html lang> sync, browser title,
  welcome-back toast (suppressed when ResumePage already showed one).
- useMenuClickOutside({langMenuRef, stepMenuRef, resumeMenuRef, ...})
  Closes the three layout dropdowns on outside click and clears the
  resume-menu copy timeout.

The residual complexity in StudyLayoutContent is JSX-driven (the
early-return waterfall — config error / status / password gate / stale
session / protected paths / completed redirect — and the desktop
stepper rendering with multiple visual states). Per the documented
exception in CLAUDE.md (precedent: AnalysisPage, StudyDesignPage,
RecruitmentPage, ConcourseDetailPage), the JSX shell carries an
explicit `// biome-ignore lint/complexity/noExcessiveCognitiveComplexity`
with rationale.

Net diff in StudyLayout.tsx: -225 / +38 (the 4 hook call sites + the
shell suppression). Frontend complexity warning count: 69 → 65 (the
StudyLayoutContent cc 74, the stepper cc 39, and the click-outside cc
16 are all gone).

Behaviour-preserving refactor: the existing StudyLayout.test.tsx and
all 985 frontend tests pass without modification.
